抽离数据库操作到 DatabaseAccessObject
