+++ src/main/java/com/github/kungua/Crawler.java
@@ -1,0 +1,98 @@
+package com.github.kungua;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import org.apache.http.HttpEntity;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.util.EntityUtils;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+
+import java.io.IOException;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.stream.Collectors;
+
+public class Crawler {
+    private DatabaseAccessObject dao = new DatabaseAccessObject();
+
+    @SuppressFBWarnings("DMI_CONSTANT_DB_PASSWORD")
+    public void run() throws Exception {
+        String link;
+//            先从数据库里拿出来一个链接(拿出来并从数据库中删除掉) 准备处理之
+        while ((link = dao.getNextLinkThenDelete()) != null) {
+            if (dao.isLinkProcessed(link)) {
+                continue;
+            }
+            if (isInterestingLink(link)) {
+                Document doc = HttpGetAndParseHtml(link);
+                parseUrlsFromPageAndStoreIntoDatabase(doc);
+                storeIntoDatabaseIfItIsNewsPage(doc, link);
+                dao.updateDatabase(link, "insert into LINKS_TO_ALREADY_PROCESSED (LINK) values (?)");
+            }
+        }
+    }
+
+    public static void main(String[] args) throws Exception {
+        new Crawler().run();
+    }
+
+    private void parseUrlsFromPageAndStoreIntoDatabase(Document doc) throws SQLException {
+        for (Element aTag : doc.select("a")) {
+            String href = aTag.attr("href");
+            if (href.startsWith("//")) {
+                href = "https:" + href;
+            }
+            if (!href.toLowerCase().startsWith("javascript")) {
+                dao.updateDatabase(href, "insert into LINKS_TO_BE_PROCESSED (LINK) values (?)");
+            }
+        }
+    }
+
+    private void storeIntoDatabaseIfItIsNewsPage(Document doc, String link) throws SQLException {
+        ArrayList<Element> articleTags = doc.select("article");
+        if (!articleTags.isEmpty()) {
+            for (Element articleTag :
+                    articleTags) {
+                String title = articleTag.child(0).text();
+                String content = articleTag.select("p").stream().map(Element::text).collect(Collectors.joining("\n"));
+                System.out.println("title = " + title);
+                dao.insertNewsIntoDatabase(link, title, content);
+            }
+        }
+    }
+
+    private static Document HttpGetAndParseHtml(String link) throws IOException {
+        CloseableHttpClient httpclient = HttpClients.createDefault();
+        System.out.println("handle link " + link);
+
+        HttpGet httpGet = new HttpGet(link);
+        httpGet.addHeader("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:68.0) Gecko/20100101 Firefox/68.0");
+
+        try (CloseableHttpResponse response = httpclient.execute(httpGet)) {
+            HttpEntity entity1 = response.getEntity();
+            String html = EntityUtils.toString(entity1);
+            return Jsoup.parse(html);
+        }
+    }
+
+    private static boolean isInterestingLink(String link) {
+        return (isNewsPage(link) || isIndexPage(link)) &&
+                isNotLoginPage(link);
+    }
+
+    private static boolean isIndexPage(String link) {
+        return "https://sina.cn".equals(link);
+    }
+
+    private static boolean isNewsPage(String link) {
+        return link.contains("news.sina.cn");
+    }
+
+    private static boolean isNotLoginPage(String link) {
+        return !link.contains("passport.sina.cn");
+    }
+}
